% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpackage21S005-package.R
\docType{package}
\name{testpackage21S005-package}
\alias{testpackage21S005}
\alias{testpackage21S005-package}
\title{testpackage21S005: Demos creating a package}
\description{
What the package does (not very much right now).
}
\author{
\strong{Maintainer}: Richard Ressler \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
